+++ src/rendering/PlacedElements.tsx
@@ -6,6 +6,7 @@ import { useUiStore } from '../state/uiStore';
 import type { SceneElement } from '../types/elements';
 import { clampToBuildArea, snapPositionXZ } from '../utilities/snap';
 import { ConveyorMesh } from './elements/ConveyorMesh';
+import { ElevatorMesh } from './elements/ElevatorMesh';
 import { intersectGround } from './groundRay';
 
 const SELECTION_COLOR = '#4f9cf0';
@@ -99,6 +100,8 @@ function PlacedElement({ element }: { element: SceneElement }) {
     >
       {element.type === 'conveyor' ? (
         <ConveyorMesh properties={element.properties} selected={selected} />
+      ) : element.type === 'elevator' ? (
+        <ElevatorMesh properties={element.properties} selected={selected} />
       ) : (
         <mesh
           position={[bounds.center.x, bounds.center.y, bounds.center.z]}